Organise features/google/src a bit more
- popUps seemed like components, thus moved to src/components
- importer.js move to importers/importers.js (like analyses/analyses.js)
