+++ features/google/src/components/popUps/index.js
@@ -1,0 +1,3 @@
+import Dialog from "../../components/dialog/dialog.jsx";
+
+export default { dialog: Dialog };
