+++ features/google/src/google.jsx
@@ -23,13 +23,13 @@ import {
     INITIAL_HISTORY_STATE,
     LoadingOverlay,
 } from "@polypoly-eu/poly-look";
-import GoogleAccount from "./model/google-account.js";
-import { dataImporters } from "./model/importer.js";
+import GoogleAccount from "./model/entities/google-account.js";
+import { importers } from "./model/importers/importers.js";
 import i18n from "!silly-i18n";
 import ExploreView from "./views/explore/explore.jsx";
 import DetailsView from "./views/explore/details.jsx";
 import ReportWrapper from "./views/report/reportWrapper.jsx";
-import BasePopUp from "./popUps/base.jsx";
+import BasePopUp from "./components/popUps/base.jsx";
 
 const Google = () => {
     const { pod, isLoading, popUp } = useContext(GoogleContext);
@@ -92,7 +92,7 @@ const GoogleApp = () => {
             <GoogleContextProvider>
                 <PolyImportProvider
                     parentContext={GoogleContext}
-                    dataImporters={dataImporters}
+                    dataImporters={importers}
                     DataAccount={GoogleAccount}
                 >
                     <div className="poly-nav-bar-separator-overlay" />
