--- features/google/src/model/importer.js
@@ -1,11 +1,0 @@
-import AccessLogImporter from "./importers/access-log-importer";
-import ActivitiesHtmlImporter from "./importers/activities-importer-html";
-import ActivitiesJsonImporter from "./importers/activities-importer-json";
-import SemanticLocationsImporter from "./importers/semantic-locations-importer";
-
-export const dataImporters = [
-    ActivitiesHtmlImporter,
-    ActivitiesJsonImporter,
-    AccessLogImporter,
-    SemanticLocationsImporter,
-];
