--- features/google/src/popUps/index.js
@@ -1,3 +1,0 @@
-import Dialog from "../components/dialog/dialog.jsx";
-
-export default { dialog: Dialog };
